refactor: flatten kernels subpackage into kernels.py
